requirements.txt
google-cloud-aiplatform[agent_engines,adk]==1.106.0
google-adk==1.9.0
httpx==0.28.1
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
cloudpickle==3.1.1
google-cloud-secret-manager==2.24.0
